+++ CHANGELOG.md
@@ -27,6 +27,7 @@ _NB at the present time, this project does not adhere strictly to Semantic Versi
 ### Fixed
 
 - Add Writing empty event list guard for `Equinox.Cosmos` [#105](https://github.com/jet/equinox/issues/105)
+- Disabled support for non-record Event payloads in `Equinox.Codec` [#103](https://github.com/jet/equinox/issues/103)
 
 <a name="1.1.0-preview2"></a>
 ## [1.1.0-preview2] - 2019-02-20